+++ src/ecommerce/index.ts
@@ -1,0 +1,2 @@
+export * from './interface';
+export * from './legacy';
