+++ src/ecommerce/interface.ts
