+++ src/index.ts
@@ -4,6 +4,7 @@ import URL from 'url-parse';
 import {
   BeaconOptions,
   BeaconTracker,
+  Displayable,
   GAOptions,
   GATracker,
   GTagOptions,
@@ -18,7 +19,6 @@ import {
   TwitterTracker,
 } from './trackers';
 import { BaseTracker, EventTracker, PageMeta } from './trackers/base';
-import { Impression } from './trackers/ecommerce/ga';
 
 export enum DeviceType {
   PC = 'pc',
@@ -182,7 +182,7 @@ export class Tracker {
   }
 
   @pushEventToQueue()
-  public sendImpression(items: Impression[]): EventParameters {
+  public sendImpression(items: Displayable[]): EventParameters {
     return [items];
   }
