+++ src/trackers/beacon/index.ts
@@ -1,0 +1,1 @@
+export * from './tracker';
