--- src/trackers/ecommerce/index.ts
@@ -1,3 +1,0 @@
-export * from './common';
-export * from './ga';
-export * from './legacy';
