+++ src/trackers/gtag.ts
@@ -1,6 +1,6 @@
 import { loadGTag } from '../utils/externalServices';
 import { BaseTracker, PageMeta } from './base';
-import { Archiveable, Displayable, Impression, Purchasable } from './ecommerce';
+import { Archiveable, Displayable, Purchasable } from '../ecommerce';
 
 export interface GTagOptions {
   trackingId: string;
