+++ src/trackers/index.ts
@@ -1,8 +1,8 @@
-export * from './beacon';
-export * from './ga';
+export * from './beacon/tracker';
+export * from './ga/tracker';
 export * from './pixel';
 export * from './tagmanager';
 export * from './gtag';
 export * from './kakao';
 export * from './twitter';
-export * from './ecommerce';
+export * from '../ecommerce';
