+++ src/trackers/pixel.ts
@@ -1,6 +1,6 @@
 import { loadPixel } from '../utils/externalServices';
 import { BaseTracker, PageMeta } from './base';
-import { Archiveable, Displayable, Impression, Purchasable } from './ecommerce';
+import { Archiveable, Displayable, Purchasable } from '../ecommerce';
 
 export interface PixelOptions {
   pixelId: string | string[];
